I18n added

+++ app/controllers/administration_controller.rb
@@ -12,14 +12,14 @@ class AdministrationController < ApplicationController
 
   def update
     com = params[:commit]
-    if com == 'Edit'
+    if com == 'Edit' || com == 'Изменить'
       @user = User.find_by(id: params[:id])
       if_edit(params[:id], params[:role])
-    elsif com == 'Delete'
+    elsif com == 'Delete' || com == 'Удалить'
       if_delete(params[:id])
-    elsif com == 'Create'
+    elsif com == 'Create' || com == 'Создать'
       redirect_to creatinguser_new_path
-    elsif com == 'Edit Subject'
+    elsif com == 'Edit Subject' || com == 'Изменить Предмет'
       redirect_to ed_subject_edit_path
     end
   end
@@ -29,24 +29,26 @@ class AdministrationController < ApplicationController
   def if_edit(id, role)
     if !User.find_by(id: id).nil? && User.find_by(id: id).role != 'admin'
       user = User.find_by(id: id)
-      redirect_to editrole_new_path(role: role, id: id), notice: "Your #{user.email} has role: #{user.role}"
+      redirect_to editrole_new_path(role: role, id: id), notice: "#{user.email} #{t('administration.ne.notice1')} #{user.role}"
     else
-      redirect_to administration_path, alert: "Invalid id"
+      redirect_to administration_path, alert: "#{t('administration.ne.error1')}"
     end
   end
 
   def if_delete(id)
     if !User.find_by(id: id).nil? && User.find_by(id: id).role != 'admin'
       user = User.find_by(id: id)
-      redirect_to administration_path, notice: "Sucsessfully deleted"
+      redirect_to administration_path, notice: "#{t('administration.ne.notice2')}"
       if user.teacher?
-        Schedule.find_by(predmet_id: Predmet.find_by(teach_id: user.teach.first.id).id).destroy unless Schedule.find_by(predmet_id: Predmet.find_by(teach_id: user.teach.first.id).id).nil?
-        List.where(predmet_id: user.teach.first.predmet.id).destroy_all unless List.find_by(predmet_id: user.teach.first.predmet.id).nil?
-        Predmet.find_by(teach_id: user.teach.first.id).destroy unless Predmet.find_by(teach_id:user.teach.first.id).nil?
+        unless Predmet.find_by(teach_id: user.teach.first.id).nil?
+          Schedule.find_by(predmet_id: Predmet.find_by(teach_id: user.teach.first.id).id).destroy unless Schedule.find_by(predmet_id: Predmet.find_by(teach_id: user.teach.first.id).id).nil?
+          List.where(predmet_id: user.teach.first.predmet.id).destroy_all unless List.find_by(predmet_id: user.teach.first.predmet.id).nil?
+          Predmet.find_by(teach_id: user.teach.first.id).destroy
+        end
       end
         user.destroy
     else
-      redirect_to administration_path, alert: "Invalid id"
+      redirect_to administration_path, alert: "#{t('administration.ne.error1')}"
     end
   end
